+++ tests/jax/test_jax_jit_qnode.py
