+++ tests/tape/test_qscript.py
